pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "awesome-curator"
version = "0.1.0"
description = "Automatically curates and updates Awesome lists from GitHub trending repositories"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["github", "awesome-list", "curation", "automation", "llm", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Internet",
]
dependencies = [
    "PyGithub>=2.1.1",
    "aiohttp>=3.9.0",
    "Jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "aiofiles>=23.2.1",
    "markdown>=3.5.0",
    "weasyprint>=60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "aioresponses>=0.7.4",
    "ruff>=0.1.0",
]

[project.scripts]
awesome-curator = "curator.__main__:main"

[project.urls]
Homepage = "https://github.com/your-org/awesome-curator"
Repository = "https://github.com/your-org/awesome-curator"
Issues = "https://github.com/your-org/awesome-curator/issues"

[tool.hatch.build.targets.wheel]
packages = ["curator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=curator --cov-report=term-missing"

[tool.coverage.run]
source = ["curator"]
omit = ["tests/*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
